{{ cookiecutter.project_slug }}/python_requirements.txt
dj-database-url
psycopg2-binary
django-webpack-loader
django-storages[s3]
whitenoise
sentry-sdk
pytest
